refactor: move etl pipeline
