+++ etl/etl_pipeline/protocols/load.py
@@ -1,6 +1,6 @@
 import logging
 
-from src.backend.db.mongodb_handler import get_mongo_collection
+from ..mongodb_handler import get_mongo_collection
 
 
 def load_protocols(protocols_data):
